requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
jinja2==3.1.2
python-dotenv==1.0.0
# Optional MySQL support
mysql-connector-python==8.2.0
